Fix collision freeze, bonus object freeze, no audio, and road border glitch

+++ source/roaddraw.c
@@ -280,7 +280,7 @@ Ptr DrawBorderZoomed16(UInt16 *drawPos,int xDrawStart,int x1,int x2,int y,UInt16
 	UInt32 u=0;
 	if(x2<0) return drawPos;
 	if(x1<0){
-		u=-x1*256*zoom;
+		u=(UInt32)(-x1*256.0f*zoom);
 		x1=0;
 	}
 	if(x2>gXSize)x2=gXSize;
